apps/backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
langchain>=0.2.0
httpx>=0.27.0
websockets>=12.0
loguru>=0.7.2
